Retry Taps in a Circular Pattern
To support hot probing, the location where the tap failed needs to be maked as fouled and not re-probed. This change adds support for remembering tap locations across a ProbeSession and substituting nearby clean locations when one becomes fouled. The distance from the original point is limited to a circle with a 4mm radius.

This also adds a choice parameter, `BAD_TAP_STRATEGY`, so any desired probe error handling behaviour can be activated on any probing command:
* `QUAD_GANTRY_LEVEL BAD_TAP_STRATEGY=RETRY`
* `BED_MESH_CALIBRATE BAD_TAP_STRATEGY=CIRCLE`

Signed-off-by: Gareth Farrington <gareth@waves.ky>